add header inside index.html and add login and signup request to the back
