+++ srcs/back/user_mgmt/login/templates/login_form.html
@@ -1,6 +1,6 @@
 <div id="login-container">
 	<div class="container">
-		<h1>LOG IN</h1>
+		<h1>{{log_in}}</h1>
 		<form id="login-form" method="post" action="http://localhost:8000/api/login/">
 			<div style="display: flex; align-items: center; justify-content: space-between;">
 				<label for="username">{{username}}:</label>
@@ -11,8 +11,9 @@
 				<input type="password" id="password" name="password">
 			</div>
 			<button id="login-submit" type="submit" value="submit">{{submit}}</button>
-			<p id="signin" style="margin: 7px;">{{dont_have_account}} <a href="#"  id="signup-link">{{click_here}}</a></p>
+			<p id="signin" style="margin: 7px;">{{dont_have_account}}<a href="/signup" data-route="/signup" id="signup-link">{{click_here}}</a></p>
 		</form>
 	</div>
-	<button id="login_intra_button" style="background-color: rgb(91, 91, 91); margin-top: 7px;">{{log_in}} 42</button>
-</div>
+
+	<button id="login_intra_button" style="background-color: rgb(91, 91, 91); margin-top: 7px;" data-route="/login-intra">{{log_in}} 42</button>
+</div>
