+++ srcs/back/user_mgmt/user_mgmt/templates/home.html
@@ -1,45 +1,46 @@
-<!-- Navbar -->
-<nav class="navbar navbar-expand-lg navbar-dark">
-    <div class="container-fluid">
-        <a class="navbar-brand" >Transcendence</a>
-        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarNav">
-            <ul class="navbar-nav ms-auto">
-                <li class="nav-item">
-                    <a id="back-to-profile-button" class="btn btn-outline-light" >Go to Profile</a>
-                </li>
-            </ul>
+<div id="profile-settings-container">
+    <!-- top navigation -->
+	<div class="d-flex justify-content-between align-items-center mb-4">
+        <h1>{{settings}}</h1>
+        <div>
+            <button data-route="/profile" id="back-to-profile-button">{{my_profile}}</button>
         </div>
     </div>
-</nav>
-
-<!-- Hero Section -->
-<section class="hero-section container-fluid">
-    <h1>Welcome to Transcendence</h1>
-    <p>mi pasion es el diseño gráfico</p>
-</section>
-
-<!-- Buttons Section -->
-<section class="container-fluid py-5">
-    <div class="row g-4 justify-content-center text-center">
-        <div class="col-md-6 col-lg-3">
-            <button class="btn btn-primary btn-large" >Play Against AI</button>
+    <form method="POST" enctype="multipart/form-data" id="profile-settings-form">
+        <div>
+            <label for="photo">{{new_photo}}:</label>
+            <input type="file" id="photo" name="photo">
         </div>
-        <div class="col-md-6 col-lg-3">
-            <button class="btn btn-success btn-large" >Play Locally</button>
+        <div>
+            <label for="username">{{username}}:</label>
+            <input type="text" id="username" name="username" value="{{ user.username }}">
         </div>
-        <div class="col-md-6 col-lg-3">
-            <button class="btn btn-warning btn-large" >Play Online</button>
+        <div>
+            <label for="first_name">{{first_name}}:</label>
+            <input type="text" id="first_name" name="first_name" value="{{ user.first_name }}">
+        </div>  
+        <div> 
+            <label for="last_name">{{last_name}}:</label>
+            <input type="text" id="last_name" name="last_name" value="{{ user.last_name }}">
         </div>
-        <div class="col-md-6 col-lg-3">
-            <button class="btn btn-danger btn-large" >Tournaments</button>
+        <div>
+            <label for="email">{{email}}:</label>
+            <input type="text" id="email" name="email" value="{{ user.email }}" readonly>
         </div>
-    </div>
-</section>
-
-<!-- Footer -->
-<footer class="footer container-fluid">
-    <p>© 2025 Transcendence. All rights reserved.</p>
-</footer>
+        <div>
+            <label for="2fa-toggle">Two-Factor Authentication:</label>
+            <input type="checkbox" id="2fa-toggle" {% if user.profile.two_fa %}checked{% endif %}>
+            <div id="2fa-setup-container" style="display: none;">
+                <p>Scan the QR code below using an authenticator app:</p>
+                <img id="2fa-qr-code" src="" alt="QR Code">
+                <p>Enter the code generated by your authenticator app:</p>
+                <input type="text" id="2fa-code" placeholder="Verification Code">
+                <button id="verify-2fa-button">Verify & Enable</button>
+            </div>
+        </div>
+        
+        <div>
+            <button type="submit" id="save-settings-button" >{{save}}</button>
+        </div>
+    </form>
+</div>
