+++ srcs/back/user_mgmt/user_profile/templates/match_history.html
@@ -2,7 +2,7 @@
     <div class="d-flex justify-content-between align-items-center mb-4">
         <h1>Match History</h1>
         <div>
-            <button id="back-to-profile-button">My profile</button>
+            <button data-route="/profile" id="back-to-profile-button">My profile</button>
         </div>
     </div>
