+++ srcs/back/user_mgmt/user_profile/templates/profile.html
@@ -4,8 +4,9 @@
         <h1>{{profile}}</h1>
         <div>
             <!-- <button id="profile-home-button">Home</button> -->
-            <button id="profile-settings-button">{{settings}}</button>
-            <button id="friends-button">{{friends}}</button>
+            <button data-route="/profile-settings" id="profile-settings-button">{{settings}}</button>
+            <button data-route="/friends" id="friends-button">{{friends}}</button>
+            <button data-route="/home" id="home-button">Home</button>
         </div>
     </div>
 
@@ -32,12 +33,12 @@
             <li>{{tournaments_played}}: <span id="tournaments-played">{{ stats_tournaments.tournaments_played }}</span></li>
             <li>{{tournament_average_score}}: <span id="tournaments-won">{{ stats_tournaments.tournament_score }} %</span></li>   
         </ul>
-        <button id="match-history-button">Match History</button>        
+        <button data-route="/match-history" id="match-history-button">Match History</button>        
     </div>
 
     <!-- Last 10 Games Chart -->
 
 
-	<button id="logout-button">Logout</button>
+	<button data-route="/logout" id="logout-button">Logout</button>
 	<button id="game">Play!</button>
 </div>
