+++ srcs/back/user_mgmt/user_profile/templates/settings_profile.html
@@ -3,7 +3,7 @@
 	<div class="d-flex justify-content-between align-items-center mb-4">
         <h1>{{settings}}</h1>
         <div>
-            <button id="back-to-profile-button">{{my_profile}}</button>
+            <button data-route="/profile" id="back-to-profile-button">{{my_profile}}</button>
         </div>
     </div>
     <form method="POST" enctype="multipart/form-data" id="profile-settings-form">
