--- srcs/back/user_mgmt/user_profile/translations.py
@@ -1,163 +1,0 @@
-from django.utils.translation import activate
-
-def add_language_context(request, context):
-	lang = get_language(request)
-	lang_words = get_lang_words(lang)
-	context.update(lang_words)
-
-def get_language(request):
-	lang = request.COOKIES.get('language') or 'en'
-	activate(lang)  # Set the current language
-	return lang
-
-def get_lang_words(lang):
-	if lang == 'es':
-		return {
-			'profile': "Perfil",
-			'welcome': "Bienvenido",
-			'settings': "Configuraciones",
-			'friends': "Amigos",
-			'username': "Nombre de usuario",
-			'name': "Nombre",
-			'my_profile': "Mi perfil",
-			'search': "Buscar",
-			'remove_friend': "Eliminar amigo",
-			'add_friend': "Agregar amigo",
-			'no_users_found': "No se encontraron usuarios",
-			'your_friends': "Tus amigos",
-			'new_photo': "Nueva foto",
-			'first_name': "Nombre",
-			'last_name': "Apellido",
-			'email': "Correo electrónico",
-			'save': "Guardar",
-			'password': "Contraseña",
-			'repeat_password': "Repetir contraseña",
-			'click_here': "Haz clic aquí",
-			'log_in': "INICIAR SESIÓN",
-			'dont_have_account': "¿No tienes una cuenta todavía?",
-			'statistics': "Estadísticas",
-			'played': "Jugados",
-			'games_won': "Ganados",
-			'tournaments_played': "Torneos jugados",
-			'tournament_average_score': "Promedio del torneo",
-		}
-	elif lang == 'ca':
-		return {
-			'profile': "Perfil",
-			'welcome': "Benvingut",
-			'settings': "Configuracions",
-			'friends': "Amics",
-			'username': "Nom d'usuari",
-			'name': "Nom",
-			'my_profile': "El meu perfil",
-			'search': "Cercar",
-			'remove_friend': "Eliminar amic",
-			'add_friend': "Afegir amic",
-			'no_users_found': "No s'han trobat usuaris",
-			'your_friends': "Els teus amics",
-			'new_photo': "Nova foto",
-			'first_name': "Nom",
-			'last_name': "Cognom",
-			'email': "Correu electrònic",
-			'save': "Desar",
-			'password': "Contrasenya",
-			'repeat_password': "Repetir contrasenya",
-			'click_here': "Fes clic aquí",
-			'log_in': "INICIAR SESSIÓ",
-			'dont_have_account': "Encara no tens un compte?",
-			'statistics': "Estadístiques",
-			'played': "Jugats",
-			'games_won': "Guanyats",
-			'tournaments_played': "Torneigs jugats",
-			'tournament_average_score': "Mitjana del torneig",
-		}
-	elif lang == 'ru':
-		return {
-			'profile': "Профиль",
-			'welcome': "Добро пожаловать",
-			'settings': "Настройки",
-			'friends': "Друзья",
-			'username': "Имя пользователя",
-			'name': "Имя",
-			'my_profile': "Мой профиль",
-			'search': "Поиск",
-			'remove_friend': "Удалить друга",
-			'add_friend': "Добавить друга",
-			'no_users_found': "Пользователи не найдены",
-			'your_friends': "Ваши друзья",
-			'new_photo': "Новое фото",
-			'first_name': "Имя",
-			'last_name': "Фамилия",
-			'email': "Электронная почта",
-			'save': "Сохранить",
-			'password': "Пароль",
-			'repeat_password': "Повторите пароль",
-			'click_here': "Нажмите здесь",
-			'log_in': "ВОЙТИ",
-			'dont_have_account': "У вас еще нет учетной записи?",
-			'statistics': "Статистика",
-			'played': "Сыграно",
-			'games_won': "Выиграно",
-			'tournaments_played': "Турниры сыграны",
-			'tournament_average_score': "Средний результат турнира",
-		}
-	elif lang == 'lv':
-		return {
-			'profile': "Profils",
-			'welcome': "Laipni lūdzam",
-			'settings': "Iestatījumi",
-			'friends': "Draugi",
-			'username': "Lietotājvārds",
-			'name': "Vārds",
-			'my_profile': "Mans profils",
-			'search': "Meklēt",
-			'remove_friend': "Noņemt draugu",
-			'add_friend': "Pievienot draugu",
-			'no_users_found': "Lietotāji nav atrasti",
-			'your_friends': "Jūsu draugi",
-			'new_photo': "Jauna fotogrāfija",
-			'first_name': "Vārds",
-			'last_name': "Uzvārds",
-			'email': "E-pasts",
-			'save': "Saglabāt",
-			'password': "Parole",
-			'repeat_password': "Atkārtot paroli",
-			'click_here': "Klikšķiniet šeit",
-			'log_in': "PIESLĒGTIES",
-			'dont_have_account': "Vai jums vēl nav konta?",
-			'statistics': "Statistika",
-			'played': "Spēlēts",
-			'games_won': "Uzvarēts",
-			'tournaments_played': "Turnīri spēlēti",
-			'tournament_average_score': "Vidējais rezultāts turnīrā",
-		}
-	else:  # Default to English if no match
-		return {
-			'profile': "Profile",
-			'welcome': "Welcome",
-			'settings': "Settings",
-			'friends': "Friends",
-			'username': "Username",
-			'name': "Name",
-			'my_profile': "My Profile",
-			'search': "Search",
-			'remove_friend': "Remove Friend",
-			'add_friend': "Add Friend",
-			'no_users_found': "No users found",
-			'your_friends': "Your Friends",
-			'new_photo': "New Photo",
-			'first_name': "First Name",
-			'last_name': "Last Name",
-			'email': "Email",
-			'save': "Save",
-			'password': "Password",
-			'repeat_password': "Repeat Password",
-			'click_here': "Click Here",
-			'log_in': "LOG IN",
-			'dont_have_account': "Don't have an account yet?",
-			'statistics': "Statistics",
-			'played': "Played",
-			'games_won': "Games Won",
-			'tournaments_played': "Tournaments Played",
-			'tournament_average_score': "Tournament Average Score",
-		}
