--- srcs/front/static/html/header.html
@@ -1,28 +1,0 @@
-<div class="position-absolute top-0 end-0 mt-4 me-4">
-    <button 
-        id="language-button" 
-        class="bg-grey border-0 text-white p-2 rounded" 
-        style="font-size: 16px; outline: none; background-color: black;">
-        EN
-    </button>
-    <ul 
-        id="language-menu"
-        class="list-unstyled bg-grey text-white p-2 rounded d-none position-absolute" 
-        style="top: 50px; right: 0; z-index: 1000;">
-        <li class="border-bottom">
-            <button class="btn btn-link text-white  text-decoration-none d-block p-2" data-lang="ca">CA</button>
-        </li>
-        <li class="border-bottom">
-            <button class="btn btn-link text-white  text-decoration-none d-block p-2" data-lang="en">EN</button>
-        </li>
-        <li class="border-bottom">
-            <button class="btn btn-link text-white text-decoration-none p-2" data-lang="es">ES</button>
-        </li>
-        <li class="border-bottom">
-            <button class="btn btn-link text-white text-decoration-none p-2" data-lang="lv">LV</button>
-        </li>
-        <li class="border-bottom">
-            <button class="btn btn-link text-white text-decoration-none p-2" data-lang="ru">RU</button>
-        </li>
-    </ul>
-</div>
