--- srcs/front/static/html/login_form.html
@@ -1,19 +1,0 @@
-<div id="login-container">
-	<div class="container">
-		<h1>LOG IN</h1>
-		<form id="login-form" method="post" action="http://localhost:8000/api/login/">
-			<div style="display: flex; align-items: center; justify-content: space-between;">
-				<label for="username">Nick:</label>
-				<input style="margin-left: 39px;" type="text" id="username" name="username">
-			</div>
-			<div style="display: flex; align-items: center; justify-content: space-between;">
-				<label for="password">Password:</label>
-				<input type="password" id="password" name="password">
-			</div>
-			<button id="login-submit" type="submit" value="submit">Submit</button>
-			<p id="signin" style="margin: 7px;">Dont have and account yet? <a href="/signup" data-route="/signup" id="signup-link">Click here</a></p>
-		</form>
-	</div>
-
-	<button id="login_intra_button" style="background-color: rgb(91, 91, 91); margin-top: 7px;" data-route="/login-intra">LOG IN 42</button>
-</div>
