--- srcs/front/static/html/signup_form.html
@@ -1,28 +1,0 @@
-<div id="login-container">
-	<div class="container">
-		<h1>Sign Up</h1>
-		<form id="signup-form" method="post" action="http://localhost:8000/api/register/">
-			<div style="display: flex; align-items: center; justify-content: space-between;">
-				<label for="username">Nick:</label>
-				<input style="margin-left: 39px;" type="text" id="username" name="username">
-			</div>
-			<div style="display: flex; align-items: center; justify-content: space-between;">
-				<label for="name">Name:</label>
-				<input style="margin-left: 39px;" type="text" id="name" name="name">
-			</div>
-			<div style="display: flex; align-items: center; justify-content: space-between;">
-				<label for="password">Password:</label>
-				<input type="password" id="password" name="password">
-			</div>
-			<div style="display: flex; align-items: center; justify-content: space-between;">
-				<label for="repassword">Repeat Password:</label>
-				<input type="password" id="repassword" name="repassword">
-			</div>
-			<div style="display: flex; align-items: center; justify-content: space-between;">
-				<label for="email">Email:</label>
-				<input type="text" id="email" name="email">
-			</div>
-			<button id="signup-submit" type="submit" value="submit">Submit</button>
-		</form>
-	</div>
-</div>
