+++ srcs/front/static/js/login.js
@@ -1,4 +1,3 @@
-
 import { loadProfilePage } from "./profile.js";
 import { handleLogout } from "./logout.js"
 import { navigateTo } from "./main.js";
@@ -81,16 +80,35 @@ export const makeAuthenticatedRequest = (url, options = {}) => {
     });
 };
 
-export const loadLoginPage = (contentArea) => {
-    fetch('html/login_form.html')  // Call the API endpoint to get the form as JSON
-        .then(response => response.text())
-        .then(data => {
-            if (data) {
-                    console.log('Form html returned!');
-                    contentArea.innerHTML = data;
-            }
-        })
-        .catch(error => console.error('Error loading login form:', error));
+// export const loadLoginPage = () => {
+//     const contentArea = document.getElementById("content-area");
+//     fetch('html/login_form.html')  // Call the API endpoint to get the form as JSON
+//         .then(response => response.text())
+//         .then(data => {
+//             if (data) {
+//                     console.log('Form html returned!');
+//                     contentArea.innerHTML = data;
+//             }
+//         })
+//         .catch(error => console.error('Error loading login form:', error));
+// };
+
+export const loadLoginPage = () => {
+    fetch(baseUrl + ":8000/api/login-form/", {
+        method: 'GET'
+    })
+    .then((response) => response.json())
+    .then(data => {
+        if (data.form_html) {
+            console.log('Form html returned!');
+            document.getElementById('content-area').innerHTML = data.form_html;
+        } else {
+            console.error('Form HTML not found in response:', data);
+        }
+    })
+    .catch(error => {
+        console.error('Error loading page', error);
+    });
 };
 
 const handleLogin = () => {
@@ -112,7 +130,8 @@ const handleLogin = () => {
             updateLanguage();
             // loadProfilePage(); //navigateTo later instead
             // navigateTo('/profile'); // change to navigate to home
-            loadHomePage();
+            // loadHomePage();
+            navigateTo('/home', true);
 
         } else {
             displayLoginError('Invalid credentials. Please try again.', 'login-form');
@@ -125,41 +144,98 @@ const handleLogin = () => {
 };
 
 export const handleSignup = () => {
-    const contentArea = document.getElementById("content-area");
     const loginForm = document.getElementById('login-form');
-    if (loginForm)
-        loginForm.remove();
-    fetch('html/signup_form.html')
-        .then(response => response.text())
-        .then(html => {
-            contentArea.innerHTML = html;
+    
+    if (loginForm) loginForm.remove();
+    
+    fetch(baseUrl + ":8000/api/signup-form/", {
+        method: 'GET',
+        headers: { 'Content-Type': 'application/json' }
+    })
+    .then(response => response.json()) // Expecting JSON response
+    .then(data => {
+        if (data && data.form_html) {
+            console.log('Form HTML returned!');
+            document.getElementById('content-area').innerHTML = data.form_html;
+
             const signupForm = document.getElementById('signup-form');
             if (signupForm) {
+                console.log('Inside if signup form!');
                 signupForm.addEventListener('submit', (event) => {
                     event.preventDefault();
-                    console.log('Submit of sign up clicked!');
+                    console.log('Submit of signup clicked!');
+                    
                     const formData = new FormData(signupForm);
-                    fetch(signupForm.action, {
+                    const formAction = signupForm.action || `${baseUrl}/api/signup/`; // Fallback action URL
+                    
+                    fetch(formAction, {
                         method: 'POST',
                         body: JSON.stringify(Object.fromEntries(formData)),
-                        headers: {'Content-Type': 'application/json'}
+                        headers: { 'Content-Type': 'application/json' }
                     })
                     .then(response => response.json())
                     .then(data => {
                         if (data.success) {
                             localStorage.setItem('access_token', data.tokens.access);
                             localStorage.setItem('refresh_token', data.tokens.refresh);
-                            loadProfilePage();
-                          // NAVIGATE TO
+                            //loadProfilePage();
+                            navigateTo('/home', true);
                         } else {
-                            displayLoginError(data.error + ' Please try again.', 'signup-form');
+                            displayLoginError(`${data.error} Please try again.`, 'signup-form');
                         }
                     })
-                })
+                    .catch(error => {
+                        console.error('Error submitting signup form:', error);
+                    });
+                });
+            } else {
+                console.error('Signup form not found in the loaded HTML.');
             }
-        })
-        .catch(error => console.error('Error loading Sign In form:', error));
+        } else {
+            console.error('Form HTML not found in response:', data);
+        }
+    })
+    .catch(error => {
+        console.error('Error loading Sign In form:', error);
+    });
 };
+
+// export const handleSignup = () => {
+//     const contentArea = document.getElementById("content-area");
+//     const loginForm = document.getElementById('login-form');
+//     if (loginForm)
+//         loginForm.remove();
+//     fetch('html/signup_form.html')
+//         .then(response => response.text())
+//         .then(html => {
+//             contentArea.innerHTML = html;
+//             const signupForm = document.getElementById('signup-form');
+//             if (signupForm) {
+//                 console.log('Exists signup form');
+//                 signupForm.addEventListener('submit', (event) => {
+//                     event.preventDefault();
+//                     console.log('Submit of sign up clicked!');
+//                     const formData = new FormData(signupForm);
+//                     fetch(signupForm.action, {
+//                         method: 'POST',
+//                         body: JSON.stringify(Object.fromEntries(formData)),
+//                         headers: {'Content-Type': 'application/json'}
+//                     })
+//                     .then(response => response.json())
+//                     .then(data => {
+//                         if (data.success) {
+//                             localStorage.setItem('access_token', data.tokens.access);
+//                             localStorage.setItem('refresh_token', data.tokens.refresh);
+//                             navigateTo('/home', true);
+//                         } else {
+//                             displayLoginError(data.error + ' Please try again.', 'signup-form');
+//                         }
+//                     })
+//                 })
+//             }
+//         })
+//         .catch(error => console.error('Error loading Sign In form:', error));
+// };
 
 export const displayLoginError = (message, form) => {
     const loginContainer = document.getElementById('login-container');
@@ -206,4 +282,3 @@ document.addEventListener("DOMContentLoaded", () => {
     //     }
     // });  
 });
-
